Add Filter Functionality for Projects
